% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{BlockedResource}
\alias{BlockedResource}
\title{BlockedResource Object}
\usage{
BlockedResource(url = NULL)
}
\arguments{
\item{url}{URL of the blocked resource}
}
\value{
BlockedResource object
}
\description{
BlockedResource Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Blocked resource.
}
