% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(mimeType = NULL, data = NULL)
}
\arguments{
\item{mimeType}{The mime-type of the image data}

\item{data}{Image data in format determined by the mime type}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describe image data.
}
