% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{MobileFriendlyIssue}
\alias{MobileFriendlyIssue}
\title{Google Search Console URL Testing Tools API Objects 
Provides tools for running validation tests against single URLs}
\usage{
MobileFriendlyIssue(rule = NULL)
}
\arguments{
\item{rule}{Rule violated}
}
\value{
MobileFriendlyIssue object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:11:59
filename: /Users/mark/dev/R/autoGoogleAPI/googlesearchconsolev1.auto/R/searchconsole_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
MobileFriendlyIssue Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mobile-friendly issue.
}
