% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{ResourceIssue}
\alias{ResourceIssue}
\title{ResourceIssue Object}
\usage{
ResourceIssue(blockedResource = NULL)
}
\arguments{
\item{blockedResource}{Describes a blocked resource issue}
}
\value{
ResourceIssue object
}
\description{
ResourceIssue Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Information about a resource with issue.
}
