% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{RunMobileFriendlyTestRequest}
\alias{RunMobileFriendlyTestRequest}
\title{RunMobileFriendlyTestRequest Object}
\usage{
RunMobileFriendlyTestRequest(requestScreenshot = NULL, url = NULL)
}
\arguments{
\item{requestScreenshot}{Whether or not screenshot is requested}

\item{url}{URL for inspection}
}
\value{
RunMobileFriendlyTestRequest object
}
\description{
RunMobileFriendlyTestRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mobile-friendly test request.
}
\seealso{
Other RunMobileFriendlyTestRequest functions: \code{\link{urlTestingTools.mobileFriendlyTest.run}}
}
