% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{RunMobileFriendlyTestResponse}
\alias{RunMobileFriendlyTestResponse}
\title{RunMobileFriendlyTestResponse Object}
\usage{
RunMobileFriendlyTestResponse(mobileFriendliness = NULL,
  mobileFriendlyIssues = NULL, screenshot = NULL, testStatus = NULL,
  resourceIssues = NULL)
}
\arguments{
\item{mobileFriendliness}{Test verdict, whether the page is mobile friendly or not}

\item{mobileFriendlyIssues}{List of mobile-usability issues}

\item{screenshot}{Screenshot of the requested URL}

\item{testStatus}{Final state of the test, can be either complete or an error}

\item{resourceIssues}{Information about embedded resources issues}
}
\value{
RunMobileFriendlyTestResponse object
}
\description{
RunMobileFriendlyTestResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Mobile-friendly test response, including mobile-friendly issues and resourceissues.
}
