% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_functions.R
\docType{package}
\name{searchconsole_googleAuthR}
\alias{searchconsole_googleAuthR}
\alias{searchconsole_googleAuthR-package}
\title{Google Search Console URL Testing Tools API
Provides tools for running validation tests against single URLs}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:11:59
filename: /Users/mark/dev/R/autoGoogleAPI/googlesearchconsolev1.auto/R/searchconsole_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}
