% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_objects.R
\name{TestStatus}
\alias{TestStatus}
\title{TestStatus Object}
\usage{
TestStatus(status = NULL, details = NULL)
}
\arguments{
\item{status}{Status of the test}

\item{details}{Error details if applicable}
}
\value{
TestStatus object
}
\description{
TestStatus Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Final state of the test, including error details if necessary.
}
