% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchconsole_functions.R
\name{urlTestingTools.mobileFriendlyTest.run}
\alias{urlTestingTools.mobileFriendlyTest.run}
\title{Runs Mobile-Friendly Test for a given URL.}
\usage{
urlTestingTools.mobileFriendlyTest.run(RunMobileFriendlyTestRequest)
}
\arguments{
\item{RunMobileFriendlyTestRequest}{The \link{RunMobileFriendlyTestRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/webmaster-tools/search-console-api/}{Google Documentation}

Other RunMobileFriendlyTestRequest functions: \code{\link{RunMobileFriendlyTestRequest}}
}
